OMG WOW NEW LOGIN SCREEN!! SettingsFragment.kt Now used as place to store settings not write them(Crazzyyy Ideas!!!) Login Uses Extension of AppHandler.kt to get needed info for signing in.
This usage uses partially cashed data to assign correctness of admin level user which is garbage and should make a call to the database to for validation but thats hard and I hate asyc calls so I give up! Asyc calls can be done somewhere else and I'll write poor code in the meantime.

Anyway the login screen works!!

Version 8C, N1.3